solutions/03_advanced/requirements.txt
llama-index>=0.9.0
llama-index-multi-modal>=0.1.3
llama-index-vector-stores-pinecone>=0.1.3
llama-index-vector-stores-weaviate>=0.1.2
llama-index-vector-stores-qdrant>=0.1.2
llama-index-embeddings-openai>=0.1.5
llama-index-llms-openai>=0.1.5
farm-haystack>=1.20.0
sentence-transformers>=2.2.2
openai>=1.3.0
pinecone-client>=2.2.4
weaviate-client>=3.24.1
qdrant-client>=1.6.0
pillow>=10.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tiktoken>=0.5.1
matplotlib>=3.7.2
pandas>=2.0.3
numpy>=1.24.3
